+++ src/client/components/AccountEditPanel/AccountEditPanel.js
@@ -308,8 +308,14 @@ Template.AccountEditPanel.events({
 
         // on update, then... update and close
         } else {
-            // update account
-            Meteor.callAsync( 'pwix_accounts_manager_accounts_update_account', self.name, item, self.item ).then( async ( res ) => {
+            const fn = instance.AM.amInstance.get().clientUpdateFn();
+            let promise;
+            if( fn ){
+                promise = fn( item, instance.AM.amInstance.get().clientUpdateArgs());
+            } else {
+                promise = Meteor.callAsync( 'pwix_accounts_manager_accounts_update_account', self.name, item, self.item );
+            }
+            promise.then( async ( res ) => {
                 if( res ){
                     res = await updateRoles( item );
                 }