+++ ChangeLog.md
@@ -18,6 +18,7 @@
     - Fix the label computation when saving an item
     - AccountsManager.s.getBy() now use the special 'getBy' permission to handle cases where we do not have any userId
     - Name the AccountEditPanel checker to make the debug easier
+    - Define preNewFn(), postNewFn(), preUpdateFn(), postUpdateFn(), clientUpdateFn() and clientUpdateArgs() functions
 
 ### 2.0.0